% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimationCKT.classif.kNN.R
\name{CKT.adaptkNN}
\alias{CKT.adaptkNN}
\title{Local aggregation of nearest neighbor estimators
of the conditional Kendall's tau by Lepski's method.}
\usage{
CKT.adaptkNN(matrixKNN, vect_k, constantA = 1, partition = NULL, verbose = 0)
}
\arguments{
\item{matrixKNN}{the matrix of nearest neighbor-estimated conditional Kendall's taus.
Each column should correspond to a choice of number of nearest neighbors
and each row should correspond to the conditional Kendall's taus
estimated at the new point.}

\item{vect_k}{vector of the numbers of nearest neighbors
(corresponding to each column of the matrixKNN).}

\item{constantA}{a tuning parameter that controls the adaptation.
The higher, the smoother it is; while the smaller, the least smooth it is.}

\item{partition}{the number of subsets to consider;
or a vector giving the id of each observations.
If \code{NULL}, only one set is used.}

\item{verbose}{if larger than 0, prints the evolution of the distances
among the path of numbers of nearest neighbors.}
}
\value{
a list with two components
\itemize{
    \item \code{estimatedCKT} the estimated conditional Kendall's tau, a vector of
    the same size as the number of rows in \code{matrixKNN};
    \item \code{vect_k_chosen} the locally selected number of nearest neighbors,
    a vector of the same size as the number of rows in \code{matrixKNN}.
}
}
\description{
Local aggregation of nearest neighbor estimators
of the conditional Kendall's tau by Lepski's method.
}
\references{
Derumigny, A., & Fermanian, J. D. (2019).
A classification point-of-view about conditional Kendall’s tau.
Computational Statistics & Data Analysis, 135, 70-94.
(Algorithm 6)
}
\keyword{internal}
