% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimationCKT.classif.GLM.R
\name{CKT.fit.GLM}
\alias{CKT.fit.GLM}
\title{Estimation of conditional Kendall's taus by penalized GLM}
\usage{
CKT.fit.GLM(
  datasetPairs,
  designMatrix = datasetPairs[, 2:(ncol(datasetPairs) - 3), drop = FALSE],
  link = "logit",
  ...
)
}
\arguments{
\item{datasetPairs}{the matrix of pairs and corresponding values of the kernel
as provided by \code{\link{datasetPairs}}.}

\item{designMatrix}{the matrix of predictor to be used for the fitting of the model.
It should hve the same number of rows as the \code{datasetPairs}.}

\item{link}{link function, can be one of
\code{logit}, \code{probit}, \code{cloglog}, \code{cauchit}).}

\item{...}{other parameters passed to
\code{ordinalNet::\link[ordinalNet]{ordinalNet}()}}
}
\value{
the fitted GLM, an object with S3 class "ordinalNet".
}
\description{
This function fits a regression model for the
conditional Kendall's tau between two variables \eqn{X_1} and \eqn{X_2}
conditionally to some predictors Z.
More precisely, this function fits the model
\deqn{\tau_{1,2|Z} = 2 \times \Lambda( \beta_0 + \beta_1 \phi_1(Z) + ... + \beta_p \phi_p(Z) )}
for a link function \eqn{\Lambda},
and \eqn{p} real-valued functions \eqn{\phi_1, ..., \phi_p}.
}
\examples{
# We simulate from a conditional copula
set.seed(1)
N = 800
Z = rnorm(n = N, mean = 5, sd = 2)
conditionalTau = 2*plogis(-1 + 0.8*Z - 0.1*Z^2) - 1
simCopula = VineCopula::BiCopSim(N=N , family = 1,
    par = VineCopula::BiCopTau2Par(1 , conditionalTau ))
X1 = qnorm(simCopula[,1])
X2 = qnorm(simCopula[,2])

datasetP = datasetPairs(X1 = X1, X2 = X2, Z = Z, h = 0.07, cut = 0.9)
designMatrix = cbind(datasetP[,2], datasetP[,2]^2)
fitCKT_GLM <- CKT.fit.GLM(
  datasetPairs = datasetP, designMatrix = designMatrix,
  maxiterOut = 10, maxiterIn = 5)
print(coef(fitCKT_GLM))
# These are rather close to the true coefficients -1, 0.8, -0.1
# used to generate the data above.

}
\references{
Derumigny, A., & Fermanian, J. D. (2019).
A classification point-of-view about conditional Kendall’s tau.
Computational Statistics & Data Analysis, 135, 70-94.
(Algorithm 2)
}
