% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimationCKT.classif.nNet.R
\name{CKT.predict.nNets}
\alias{CKT.predict.nNets}
\title{Predict the values of conditional Kendall's tau
using Model Averaging of Neural Networks}
\usage{
CKT.predict.nNets(fit, newZ, aggregationMethod = "mean")
}
\arguments{
\item{fit}{result of a call to \code{CKT.fit.nNet}}

\item{newZ}{new matrix of observations, with the same number of variables.
and same names as the \code{designMatrix} that was used to fit the neural networks.}

\item{aggregationMethod}{the method to be used to aggregate all the predictions
together. Can be \code{"mean"} or \code{"median"}.}
}
\value{
\code{CKT.predict.nNets} returns
a vector of (predicted) conditional Kendall's taus of the same size
as the number of rows of the matrix \code{newZ}.
}
\description{
Predict the values of conditional Kendall's tau
using Model Averaging of Neural Networks
}
