% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimationCKT.classif.nNet.R
\name{CKT.fit.nNets}
\alias{CKT.fit.nNets}
\title{Estimation of conditional Kendall's taus by model averaging of Neural Networks}
\usage{
CKT.fit.nNets(
  datasetPairs,
  designMatrix = datasetPairs[, 2:(ncol(datasetPairs) - 3), drop = FALSE],
  vecSize = rep(3, times = 10),
  nObs_per_NN = 0.9 * nrow(designMatrix),
  verbose = 1
)
}
\arguments{
\item{datasetPairs}{the matrix of pairs and corresponding values of the kernel
as provided by \code{\link{datasetPairs}}.}

\item{designMatrix}{the matrix of predictor to be used for the fitting of the tree}

\item{vecSize}{vector with the number of neurons for each network}

\item{nObs_per_NN}{number of observations used for each neural network.}

\item{verbose}{a number indicated what to print
\itemize{
    \item \code{0}: nothing printed at all.
    \item \code{1}: a message is printed at the convergence of each neural network.
    \item \code{2}: each optimization of each network is printed in detail.
}}
}
\value{
a list of the fitted neural networks
}
\description{
Estimation of conditional Kendall's taus by model averaging of Neural Networks
}
\examples{
# We simulate from a conditional copula
set.seed(1)
N = 800
Z = rnorm(n = N, mean = 5, sd = 2)
conditionalTau = -0.9 + 1.8 * pnorm(Z, mean = 5, sd = 2)
simCopula = VineCopula::BiCopSim(N=N , family = 1,
    par = VineCopula::BiCopTau2Par(1 , conditionalTau ))
X1 = qnorm(simCopula[,1])
X2 = qnorm(simCopula[,2])

newZ = seq(2,10,by = 0.1)
datasetP = datasetPairs(X1 = X1, X2 = X2, Z = Z, h = 0.07, cut = 0.9)
fitCKT_nets <- CKT.fit.nNets(datasetPairs = datasetP)


}
\references{
Derumigny, A., & Fermanian, J. D. (2019).
A classification point-of-view about conditional Kendall’s tau.
Computational Statistics & Data Analysis, 135, 70-94.
(Algorithm 7)
}
