% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimationCKT.kernelCV.R
\name{CKT.hCV.Kfolds}
\alias{CKT.hCV.Kfolds}
\title{Choose the bandwidth for estimation of conditional Kendall's tau
using n-fold cross-validation}
\usage{
CKT.hCV.Kfolds(
  range_h,
  matrixSignsPairs,
  observedZ,
  ZToEstimate,
  typeEstCKT = 4,
  kernel.name = "Epa",
  Kfolds = 5,
  progressBar = TRUE
)
}
\arguments{
\item{range_h}{vector containing possible values for the bandwidth.}

\item{matrixSignsPairs}{square matrix of signs of all pairs,
produced by computeMatrixSignPairs.}

\item{observedZ}{vector of observed values of Z.
If Z is multivariate, then this is a matrix whose rows correspond
to the observations of Z}

\item{ZToEstimate}{vector of fixed conditioning values at which
the difference between the two conditional Kendall's tau should be computed.
Can also be a matrix whose lines are the conditioning vectors at which
the difference between the two conditional Kendall's tau should be computed.}

\item{typeEstCKT}{type of estimation of the conditional Kendall's tau.}

\item{kernel.name}{name of the kernel used for smoothing.
Possible choices are "Gaussian" (Gaussian kernel) and "Epa" (Epanechnikov kernel).}

\item{Kfolds}{number of subsamples used.}

\item{progressBar}{if TRUE, a progressbar for each h is displayed
to show the progress of the computation.}
}
\value{
A list with components:
\itemize{
    \item \code{hCV}: the chosen bandwidth
    \item \code{scores}: vector of the same length as range_h giving the
    value of the CV criteria for each of the h tested
}
}
\description{
Choose the bandwidth for estimation of conditional Kendall's tau
using n-fold cross-validation
}
\references{
Derumigny, A., & Fermanian, J. D. (2019).
On kernel-based estimation of conditional Kendall’s tau:
finite-distance bounds and asymptotic behavior.
Dependence Modeling, 7(1), 292-321.
}
