% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimationCKT.KendallReg.R
\name{CKT.KendallReg.LambdaCV}
\alias{CKT.KendallReg.LambdaCV}
\title{Kendall's regression: choice of the penalization parameter by K-folds cross-validation}
\usage{
CKT.KendallReg.LambdaCV(
  observedX1,
  observedX2,
  observedZ,
  ZToEstimate,
  designMatrixZ,
  typeEstCKT = 4,
  h_lambda,
  Lambda = identity,
  kernel.name = "Epa",
  Kfolds_lambda = 10,
  l_norm = 1,
  matrixSignsPairs = NULL,
  progressBars = "global"
)
}
\arguments{
\item{observedX1}{a vector of n observations of the first variable \eqn{X_1}.}

\item{observedX2}{a vector of n observations of the second variable \eqn{X_2}.}

\item{observedZ}{a vector of n observations of the conditioning variable,
or a matrix with n rows of observations of the conditioning vector
(if \eqn{Z} is multivariate).}

\item{ZToEstimate}{the new data of observations of Z at which
the conditional Kendall's tau should be estimated.}

\item{designMatrixZ}{the transformation of the ZToEstimate that
will be used as predictors. By default, no transformation is applied.}

\item{typeEstCKT}{type of estimation of the conditional Kendall's tau.}

\item{h_lambda}{the smoothing bandwidth used in the cross-validation
procedure to choose \code{lambda}.}

\item{Lambda}{the function to be applied on conditional Kendall's tau.
By default, the identity function is used.}

\item{kernel.name}{name of the kernel. Possible choices are
"Gaussian" (Gaussian kernel) and "Epa" (Epanechnikov kernel).}

\item{Kfolds_lambda}{the number of folds used in the cross-validation
procedure to choose \code{lambda}.}

\item{l_norm}{type of norm used for selection of the optimal lambda.
l_norm=1 corresponds to the sum of absolute values of differences
between predicted and estimated conditional Kendall's tau
while l_norm=2 corresponds to the sum of squares of differences.}
}
\description{
In this model, three variables \eqn{X_1}, \eqn{X_2} and \eqn{Z} are observed.
We try to model the conditional Kendall's tau between \eqn{X_1} and \eqn{X_2} conditionally
to \eqn{Z=z}, as follows:
\deqn{\Lambda(\tau_{X_1, X_2 | Z = z})
= \sum_{i=1}^{p'} \beta_i \psi_i(z),}
where \eqn{\tau_{X_1, X_2 | Z = z}} is the conditional Kendall's tau
between \eqn{X_1} and \eqn{X_2} conditionally to \eqn{Z=z},
\eqn{\Lambda} is a function from \eqn{]-1, 1[]} to \eqn{R},
\eqn{(\beta_1, \dots, \beta_p)} are unknown coefficients to be estimated
and \eqn{\psi_1, \dots, \psi_{p'})} are a dictionary of functions.
To estimate \eqn{beta}, we used the penalized estimator which is defined
as the minimizer of the following criteria
\deqn{\frac{1}{2n'} \sum_{i=1}^{n'} [\Lambda(\hat\tau_{X_1, X_2 | Z = z})
- \sum_{j=1}^{p'} \beta_j \psi_j(z)]^2 + \lambda * |\beta|_1.}
This function chooses the penalization parameter \eqn{lambda}
by cross-validation.
}
\references{
Derumigny, A., & Fermanian, J. D. (2020).
On Kendall’s regression.
Journal of Multivariate Analysis, 178, 104610.
}
\keyword{internal}
