% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimationCKT.KendallReg.R
\name{CKT.kendallReg.predict}
\alias{CKT.kendallReg.predict}
\title{Predict conditional Kendall's tau using Kendall's regression}
\usage{
CKT.kendallReg.predict(fit, newData, lambda = NULL, Lambda_inv = identity)
}
\arguments{
\item{fit}{the fitted model, obtained by a call
to \code{\link{CKT.kendallReg.fit}}}

\item{newData}{the new observations of the conditioning variable}

\item{lambda}{value or vector of values of the penalty parameters with
which to predict using the penalized linear model. By default, it uses
the sequence used to fit the model.}

\item{Lambda_inv}{the functional inverse of the transformation \eqn{Lambda}
used to fit the Kendall's regression.}
}
\value{
the predicted values
}
\description{
Uses new data of the conditioning variable Z to predict
the corresponding conditional Kendall's tau between two variables
\eqn{X_1} and \eqn{X_2} given \eqn{Z=z} using some new
values of \eqn{Z}
}
\references{
Derumigny, A., & Fermanian, J. D. (2020).
On Kendall’s regression.
Journal of Multivariate Analysis, 178, 104610.
}
