% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimationCKT.kernel.R
\name{CKT.kernel.univariate}
\alias{CKT.kernel.univariate}
\title{Estimate the conditional Kendall's tau of X1 and X2
at different points}
\usage{
CKT.kernel.univariate(
  matrixSignsPairs,
  observedZ,
  h,
  ZToEstimate,
  kernel.name = "Epa",
  typeEstCKT = 4,
  progressBar = TRUE
)
}
\arguments{
\item{matrixSignsPairs}{square matrix of signs of all pairs,
produced by computeMatrixSignPairs.}

\item{observedZ}{vector of observed points of Z.
It shall have the same length as the number of rows of \code{matrixSignsPairs}.}

\item{h}{bandwidth. It can be a real, in this case the same \code{h}
will be used for every element of \code{vectorZToEstimate}.
If \code{h}is a vector then its elements are recycled to match the length of
\code{vectorZToEstimate}.}

\item{ZToEstimate}{points at which the conditional Kendall's tau is computed.}

\item{kernel.name}{name of the kernel used for smoothing.
Possible choices are "Gaussian" (Gaussian kernel) and "Epa" (Epanechnikov kernel).}

\item{typeEstCKT}{type of estimation of the conditional Kendall's tau.}

\item{progressBar}{if TRUE, a progressbar is displayed to show the computation.}
}
\value{
a vector of the same length as vectorZToEstimate whose elements
are the estimated conditional Kendall's taus of X1 and X2 given Z = z.
}
\description{
Estimate the conditional Kendall's tau of X1 and X2
at different points
}
\keyword{internal}
