% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimationCKT.kernel.R
\name{CKT.kernelPointwise.univariate}
\alias{CKT.kernelPointwise.univariate}
\title{Estimate the conditional Kendall's tau of X1 and X2
at a fixed univariate point Z = pointZ}
\usage{
CKT.kernelPointwise.univariate(
  matrixSignsPairs,
  vectorZ,
  h,
  pointZ,
  kernel.name,
  typeEstCKT
)
}
\arguments{
\item{matrixSignsPairs}{square matrix of signs of all pairs,
produced by computeMatrixSignPairs.}

\item{vectorZ}{vector of observed points of Z.
It shall have the same length as the number of rows of matrixSignsPairs.}

\item{h}{bandwidth}

\item{pointZ}{point at which the conditional Kendall's tau is computed.}

\item{kernel.name}{name of the kernel used for smoothing.}

\item{typeEstCKT}{type of estimation of the conditional Kendall's tau.}
}
\value{
an estimator of the conditional Kendall's tau
of X1 and X2 given Z = z.
}
\description{
Estimate the conditional Kendall's tau of X1 and X2
at a fixed univariate point Z = pointZ
}
\keyword{internal}
