% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimationCKT.classif.tree.R
\name{CKT.predict.tree}
\alias{CKT.predict.tree}
\title{Predict the values of conditional Kendall's tau a classification tree}
\usage{
CKT.predict.tree(fit, newData)
}
\arguments{
\item{fit}{result of a call to CKT.fit.tree}

\item{newData}{new matrix of observations, with the same number of variables.
and same names as the designMatrix that was used to fit the tree.}
}
\value{
a vector of (predicted) conditional Kendall's taus of the same size
as the number of rows of the newData.
}
\description{
Predict the values of conditional Kendall's tau a classification tree
}
\examples{
# We simulate from a conditional copula
set.seed(1)
N = 800
Z = rnorm(n = N, mean = 5, sd = 2)
conditionalTau = -0.9 + 1.8 * pnorm(Z, mean = 5, sd = 2)
simCopula = VineCopula::BiCopSim(N=N , family = 1,
    par = VineCopula::BiCopTau2Par(1 , conditionalTau ))
X1 = qnorm(simCopula[,1])
X2 = qnorm(simCopula[,2])

datasetP = datasetPairs(X1 = X1, X2 = X2, Z = Z, h = 0.07, cut = 0.9)
est_Tree = CKT.fit.tree(datasetPairs = datasetP, mindev = 0.008)

newData = seq(1,10,by = 0.1)
prediction = CKT.predict.tree(fit = est_Tree, newData = data.frame(x=newData))
# Comparison between true Kendall's tau (in red)
# and estimated Kendall's tau (in black)
plot(newData, prediction, type = "l", ylim = c(-1,1))
lines(newData, -0.9 + 1.8 * pnorm(newData, mean = 5, sd = 2), col="red")

}
\references{
Derumigny, A., & Fermanian, J. D. (2019).
A classification point-of-view about conditional Kendall’s tau.
Computational Statistics & Data Analysis, 135, 70-94.
(Section 3.2)
}
