% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimationCKT.datasetPairs.R
\name{datasetPairs_hCV}
\alias{datasetPairs_hCV}
\title{Construct a dataset of pairs to be used for the cross-validation
choice of the bandwidth for kernel estimation of conditional Kendall's tau}
\usage{
datasetPairs_hCV(X1, X2, Z, nPairs = NULL, typeEstCKT = 2)
}
\arguments{
\item{nPairs}{number of most relevant pairs to keep in the final datasets.
If \code{NULL}, no cut is realized.}

\item{typeEstCKT}{type of estimation of the conditional Kendall's tau.}

\item{observedX1}{a vector of n observations of the first variable.}

\item{observedX2}{a vector of n observations of the second variable.}
}
\value{
the corresponding dataset of pairs
}
\description{
Construct a dataset of pairs to be used for the cross-validation
choice of the bandwidth for kernel estimation of conditional Kendall's tau
}
\keyword{internal}
