% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimationCondCopulas.R
\name{estimateParCondCopula}
\alias{estimateParCondCopula}
\title{Estimation of parametric conditional copulas}
\usage{
estimateParCondCopula(
  observedX1,
  observedX2,
  observedX3,
  newX3,
  family,
  method = "mle",
  h
)
}
\arguments{
\item{observedX1}{a vector of n observations of the first variable}

\item{observedX2}{a vector of n observations of the second variable}

\item{observedX3}{a vector of n observations of the conditioning variable}

\item{newX3}{a vector of new observations of X3}

\item{family}{an integer indicating the parametric family of copulas to be used,
following the conventions of the package \code{VineCopula}.}

\item{method}{the method of estimation of the conditional parameters.
Can be \code{"mle"} for maximum likelihood estimation
or \code{"itau"} for estimation by inversion of Kendall's tau.}

\item{h}{bandwidth to be chosen}
}
\value{
a vector of size \code{length(newX3)} containing
the estimated conditional copula parameters for each value of newU3.
}
\description{
This function computes an estimate of the conditional parameters
of a conditional parametric copula model.
}
\details{
We assume that we have a three-dimensional vector (X1, X2, X3).
We are interested in the estimation of the conditional copula of (X1, X2) given X3.
This function takes in parameter three vectors of observations of X1, X2, X3,
and a vector of conditioning points \code{newX3}
and compute estimators of the corresponding parametric conditional copulas.
}
\examples{

# We simulate from a conditional copula
N = 5000

X3 = rnorm(n = N, mean = 5, sd = 2)
conditionalTau = 0.9 * pnorm(X3, mean = 5, sd = 2)
simCopula = VineCopula::BiCopSim(
    N=N , family = 1, par = VineCopula::BiCopTau2Par(1 , conditionalTau ))
X1 = qnorm(simCopula[,1])
X2 = qnorm(simCopula[,2])

gridnewX3 = seq(2, 8, by = 1)
conditionalTauNewX3 = 0.9 * pnorm(gridnewX3, mean = 5, sd = 2)

vecEstimatedThetas = estimateParCondCopula(
  observedX1 = X1, observedX2 = X2, observedX3 = X3,
  newX3 = gridnewX3, family = 1, h = 0.1)

# Estimated conditional parameters
vecEstimatedThetas
# True conditional parameters
VineCopula::BiCopTau2Par(1 , conditionalTauNewX3 )

# Estimated conditional Kendall's tau
VineCopula::BiCopPar2Tau(1 , vecEstimatedThetas )
# True conditional Kendall's tau
conditionalTauNewX3


}
