% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimationCondCopulas.R
\name{estimateParCondCopula_ZIJ}
\alias{estimateParCondCopula_ZIJ}
\title{Estimation of a parametric conditional copula using conditional pseudos-observations}
\usage{
estimateParCondCopula_ZIJ(Z1_J, Z2_J, U3, newU3, family, method, h)
}
\arguments{
\item{Z1_J}{the conditional pseudos-observations of the first variable,
i.e. \eqn{\hat F_{1|J}( x_{i,1} | x_J = x_{i,J})} for \eqn{i=1,\dots, n}.}

\item{Z2_J}{the conditional pseudos-observations of the second variable,
i.e. \eqn{\hat F_{2|J}( x_{i,2} | x_J = x_{i,J})} for \eqn{i=1,\dots, n}.}

\item{U3}{the pseudos-observations of the conditional variable.}

\item{newU3}{new observations of the conditionning variable on which we would like
to estimate the conditional parameter}

\item{family}{the family of conditional copulas used.}

\item{method}{method of estimation for the parametric conditional copula.}

\item{h}{the bandwidth.}
}
\value{
a vector of size \code{length(newX3)} containing
the estimated conditional copula parameters for each value of newU3.
}
\description{
This is an auxiliary function that is called when conditional pseudos-observations
are already available to estimate a parametric conditional copula.
}
