% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bCond.tree.conversion.R
\name{matrixInd2matrixCKT}
\alias{matrixInd2matrixCKT}
\title{Construction of the matrix of estimated CKT given a matrix of indicators}
\usage{
matrixInd2matrixCKT(matrixInd, newDataXI)
}
\arguments{
\item{matrixInd}{a matrix of indexes of size (n, N.boxes) describing
for each observation i to which box ( = event) it belongs.}

\item{newDataXI}{this is a matrix of size \eqn{n' \times |I|}
where \eqn{|I|} is the number of conditioned variables.
By default this is \code{NULL} meaning that
we return the matrix of conditional Kendall's tau for the original data
used to compute the \code{estimatedTree}}
}
\value{
a matrix of size \eqn{|I|(|I|-1) \times m} where each component corresponds
to a conditional Kendall's tau between a pair of conditional variables
conditionally to the conditioned variables in one of the boxes
}
\description{
Construction of the matrix of estimated CKT given a matrix of indicators
}
