% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_generics.R
\name{se}
\alias{se}
\title{Standard Error of Model Coefficients}
\usage{
se(object, ...)
}
\arguments{
\item{object}{A model object.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Generic function to calculate the standard error of estimated coefficients
from a model object.
}
