% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bCond.tree.conversion.R
\name{treeCKT2matrixInd}
\alias{treeCKT2matrixInd}
\title{Construction of the matrix of indicators given an estimated CKT tree}
\usage{
treeCKT2matrixInd(estimatedTree, newDataXJ = NULL)
}
\arguments{
\item{estimatedTree}{the tree that has been estimated before,
for example by \code{\link{bCond.treeCKT}}.}

\item{newDataXJ}{this is a matrix of size \eqn{n' \times |J|}
where \eqn{|J|} is the number of conditional variables used in the tree.
By default this is \code{NULL} meaning that
we return the matrix of indicators for the original data
(that was used to compute the \code{estimatedTree}).}
}
\value{
a matrix of size \eqn{n' \times m} which component \code{[i,j]}
is \eqn{1\{ X_{i,J} \in A_{j,J} \}}.
}
\description{
Construction of the matrix of indicators given an estimated CKT tree
}
