% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimationCKT.kernelCV.R
\name{CKT.hCV.l1out}
\alias{CKT.hCV.l1out}
\title{Leave-one-out cross validation for choosing the bandwidth
for kernel-based range_h of conditional Kendall's tau}
\usage{
CKT.hCV.l1out(
  observedX1,
  observedX2,
  observedZ,
  range_h,
  matrixSignsPairs,
  nPairs = 10 * length(observedX1),
  typeEstCKT = 4,
  kernel.name = "Epa",
  progressBar = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{observedX1}{a vector of n observations of the first variable}

\item{observedX2}{a vector of n observations of the second variable}

\item{observedZ}{a vector of n observations of the conditioning variable,
or a matrix with n rows of observations of the conditioning vector}

\item{range_h}{vector containing possible values for the bandwidth.}

\item{matrixSignsPairs}{square matrix of signs of all pairs,
produced by computeMatrixSignPairs.}

\item{nPairs}{number of pairs used in the cross-validation criteria.}

\item{typeEstCKT}{type of estimation of the conditional Kendall's tau.}

\item{kernel.name}{name of the kernel used for smoothing.
Possible choices are "Gaussian" (Gaussian kernel) and "Epa" (Epanechnikov kernel).}

\item{progressBar}{if TRUE, a progressbar for each h is displayed
to show the progress of the computation.}

\item{verbose}{if TRUE, print the score of each h during the procedure.}
}
\value{
A list with components:
\itemize{
    \item \code{hCV}: the chosen bandwidth
    \item \code{scores}: vector of the same length as range_h giving the
    value of the CV criteria for each of the h tested
}
}
\description{
Leave-one-out cross validation for choosing the bandwidth
for kernel-based range_h of conditional Kendall's tau
}
\references{
Derumigny, A., & Fermanian, J. D. (2019).
On kernel-based estimation of conditional Kendall’s tau:
finite-distance bounds and asymptotic behavior.
Dependence Modeling, 7(1), 292-321.
Page 296, Equation (4).
}
