% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CKTmatrix.kernel.R
\name{CKTmatrix.kernel}
\alias{CKTmatrix.kernel}
\title{Estimate the (averaging) conditional Kendall's tau matrix
at different points}
\usage{
CKTmatrix.kernel(
  dataMatrix,
  observedZ,
  gridZ,
  typeMatrixCKT = "all",
  blockStructure = NULL,
  h,
  kernel.name = "Epa",
  typeEstCKT = 4
)
}
\arguments{
\item{dataMatrix}{n x d matrix of d-dimensional multivariate random variable
of n timepoints.}

\item{observedZ}{vector of observed points of Z.
It must have the same length as the number of rows of \code{dataMatrix}.}

\item{gridZ}{points at which the conditional Kendall's tau is computed.}

\item{typeMatrixCKT}{name of the matrix estimator used. Possible choices are
"All" (no averaging),
"AveDiag" (averaging along diagonal block elements) and
"AveAll" (averaging all CKT's within blocks)}

\item{blockStructure}{list of groups. A group is a vector with
variable numbers. \code{blockStructure} must be a partition of 1:d, where d is
the amount of columns in \code{dataMatrix}.}

\item{h}{bandwidth. It can be a real, in this case the same \code{h}
will be used for every element of \code{gridZ}.
If \code{h}is a vector then its elements are recycled to match the length of
\code{gridZ}.}

\item{kernel.name}{name of the kernel used for smoothing.
Possible choices are "Gaussian" (Gaussian kernel) and "Epa" (Epanechnikov kernel).}

\item{typeEstCKT}{type of estimation of the conditional Kendall's tau.}
}
\value{
array with dimensions depending on \code{typeMatrixCKT}.
If \code{typeMatrixCKT} = "all", it returns an array of dimension
n x n x length(\code{gridZ}), giving the conditional Kendall's tau matrix
given Z = z. Here, n is the number of rows in \code{dataMatrix}.
If \code{typeMatrixCKT} = "aveDiag" or "aveAll" the function returns an array
of dimension length(\code{blockStructure}) x
length(\code{blockStructure}) x  length(\code{gridZ}), giving the block
estimates of the conditional Kendall's tau given Z = z with ones on the
diagonal.
}
\description{
Estimate the (averaging) conditional Kendall's tau matrix
at different points
}
