% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpA.measures.R
\name{measures_nonsimplifyingness_NP}
\alias{measures_nonsimplifyingness_NP}
\title{Compute a measure of non-simplifyingness based on non-parametric estimation
of the conditional copula}
\usage{
measures_nonsimplifyingness_NP(
  X1,
  X2,
  Z,
  h,
  measures = "all",
  kernel.name = "Epanechnikov",
  truncVal = h,
  numericalInt = list(kind = "legendre", nGrid = 10)
)
}
\arguments{
\item{X1, X2}{vector of \code{n} observations of the conditioned variables}

\item{Z}{vector of \code{n} observations of the conditioning variable}
}
\description{
Compute a measure of non-simplifyingness based on non-parametric estimation
of the conditional copula
}
