% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpA.measures.R
\name{measures_nonsimplifyingness_NP}
\alias{measures_nonsimplifyingness_NP}
\title{Compute a measure of non-simplifyingness based on non-parametric estimation
of the conditional copula}
\usage{
measures_nonsimplifyingness_NP(
  X1,
  X2,
  Z,
  h,
  measures = "all",
  kernel.name = "Epanechnikov",
  truncVal = NULL,
  numericalInt = list(kind = "legendre", nGrid = 10),
  verbose = 0
)
}
\arguments{
\item{X1, X2}{vector of \code{n} observations of the conditioned variables}

\item{Z}{vector of \code{n} observations of the conditioning variable}

\item{h}{the bandwidth used for kernel smoothing}

\item{measures}{choices of measures of non-simplifyingness to be computed.
\code{measures = "all"} includes all available non-parametric measures of
non-simplifyingness. Otherwise, \code{measures} must be a character vector
and a subset of \code{c("T1_CvM_Cs3", "T1_CvM_Cs4", "tilde_T0_CvM", }
\code{"T1_KS_Cs3", "T1_KS_Cs4", "tilde_T0_KS")}.}

\item{kernel.name}{the name of the kernel}

\item{truncVal}{the value of truncation for the integral,
i.e. the integrals are computed from \code{truncVal} to \code{1-truncVal}
instead of from 0 to 1.
Note that \code{truncVal} must be in the interval \eqn{[0, 0.5)},
i.e. \eqn{0} is allowed but not \eqn{0.5}.

The default is \code{truncVal = NULL}, which actually means that
\code{truncVal = h} if \code{h < 0.5} and \code{truncVal = 0} else.}

\item{numericalInt}{parameters to be given to
\code{statmod::\link[statmod]{gauss.quad}}, including the number of
quadrature points and the type of interpolation.}

\item{verbose}{option used for debugging. If \code{verbose = 0}, the function
is silent. Higher values of \code{verbose} give more explicit details on the
computations.}
}
\description{
Compute a measure of non-simplifyingness based on non-parametric estimation
of the conditional copula
}
